entrypoint: add support for a 'watch completion file' mode
In certain cases (like various Kubernetes uses as a sidecar), it is
useful to terminate the process under certain conditions -- like if a
file is found.

If the WATCH_COMPLETION_FILE environment var is set, then if a file
named 'sim-completed' appears in the log directory, the Python code will
immediately exit instead of ingesting the next batch of log files.

The general intention is that you can use this with a simulation harness
to run a bunch of simulations, output their trace files (and thus, load
them into ClickHouse), and shut down the ingestion server by writing a
file to mark things as complete.

At the moment, this is primarily useful because Kubernetes Jobs cannot
have proper 'sidecars' that exit/are killed after some primary container
runs.

Signed-off-by: Austin Seipp <aseipp@pobox.com>