trace-converter: properly truncate milliseconds for Time column
Pandas dataframes can parse nanosecond resolution Timestamp values. The
timestamps emitted in FDB trace logs use millisecond resolution, meaning
they have a fractional component -- while ClickHouse does not support
sub-second resolution in any way. Pandas always stores this fractional
component internally for all datetime types.

Before, we were trying to get away with formatting the output using a
trick to convert to .astype('datetime[ns]') effectively saying we don't
care about sub-second timestamp resolution -- but the problem is the
DataFrame still stores the subsecond resolution internally, and when
emitting JSON records, it's still retained. This means the output value,
a numeric value representing a Unix timestamp, is scaled by milliseconds
(multiplied by 1000).

This meant that when DataFrames were emitted as JSON records to
ClickHouse, we got Time columns that looked like this:

    >>> df.head(20).to_json(orient='records', lines=True)
    {"Time":1557908434000,"Type":"Role"} ...

Note the '000' is the empty millisecond value. Truncated but still
scaled in the output. How does ClickHouse deal with this?

Well, the toDateTime function can take either integers, or strings. If
you give it a string like:

    > SELECT toDateTime('1557908434000')

    Received exception from server (version 19.5.2):
    Code: 6. DB::Exception: Received from clickhouse-server:9000, 10.1.0.3.
    DB::Exception: Cannot parse string '1557908434000' as DateTime: syntax
    error at position 10 (parsed just '1557908434').

it properly fails: ClickHouse can not handle the sub-second resolution
present in this string timestamp.

But the JSON record is _not_ a string for the 'Time' field, it's an
Int64 or whatnot. I originally thought this was submitted as a string,
but alas. And the semantics of toDateTime change here:

    > SELECT toDateTime(1557908434000)

    ┌─toDateTime(1557908434000)─┐
    │       2069-03-12 00:14:08 │
    └───────────────────────────┘

Which parses this integer and puts it somewhere into the future.

Luckily this can all be avoided easily: simply cast the Time column to
an 'int', and it throws away the sub-second resolution for you!

    >>> df.head(20).astype(int).to_json(orient='records', lines=True)
    {"Time":1557908434,"Type":"Role"} ...

Which gives the proper result:

    > SELECT toDateTime(1557908434)

    ┌─toDateTime(1557908434)─┐
    │    2019-05-15 08:20:34 │
    └────────────────────────┘

And all is well.

Signed-off-by: Austin Seipp <aseipp@pobox.com>